pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-customer-service-agent"
version = "1.0.0"
description = "A production-ready AI customer service agent system"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
notebooks = ["jupyter>=1.0.0", "ipykernel>=6.0.0"]

[project.urls]
Homepage = "https://github.com/Bhavik-Jikadara/ai-customer-service-agent.git"
Documentation = "https://github.com/Bhavik-Jikadara/ai-customer-service-agent.git/docs"
Repository = "https://github.com/Bhavik-Jikadara/ai-customer-service-agent.git"
Issues = "https://github.com/Bhavik-Jikadara/ai-customer-service-agent.git/issues"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
